Use friso prefix for --with-friso
Because groonga-tokenizer-friso requires both Friso's lib dir and
include dir. It is better that we receive prefix of them and generates
lib dir and include dir from the prefix rather than guessing include
dir from lib dir.

Or we should receive both lib dir and include dir by options like
`--with-friso-lib-dir` and `--with-friso-include-dir`.